(+) Very first commit of surface.dll project, an extension module that draw surfaces of planets. The original code is from gltest, so it's C based. It was tough to just make the codes in C++ manners. Of course, actual drawing is very crappy yet.
I could make this a branch, but doing so would unnecessarily complicate revision tree.
it's not much trouble maintaining this module though if it's not used, because extensions can be completely excluded from the execution at any time. Rather, the main program is more desirable if the interface is more usable by many extensions.

git-svn-id: svn://localhost/gltestplus/branches/bullet@444 42c370b9-4569-d348-85c3-ab3aefc53057