(+) Added Bullet Dynamics Engine support to WarMaps. It looks not quite right yet, but at least we can keep objects on (near) the ground.
I don't really have an idea how to link Bullet library to the application program and to the extensions.
Right now, Bullet statically links to application or extension, depending on where the Bullet class object is allocated.
For example, if we 'new' btRigidBody in gltestplus.exe, it'll use virtual function implementations linked into gltestplus.exe. On the other hand, if we do it in surface.dll, it'll link to the dll.
I cannot figure how to avoid duplicate function definition with VC's linking specification but to make Bullet separate DLL, and I don't like so many unnecessary DLLs. Besides, Bullet does not export it's member classes, so we must modify Bullet's sources.
But the fact is that no problem exposes yet. As long as the allocator always does deallocation too, no memory-manager-involved problems will occur. This is still very bad design, though.


git-svn-id: svn://localhost/gltestplus/branches/bullet@453 42c370b9-4569-d348-85c3-ab3aefc53057