fixes #34 Added server command system to notify how the client wants to interact with the server.
The Player's rotation quaternion is the first case of this interaction. Now the server Player and client Player can move their eyes without interfering each other.

Also, added serverrecv.log to save messages sent from the client and received by the server for debugging.

The client is very slow for some reason, but it's another story.


git-svn-id: svn://localhost/gltestplus/branches/bullet@588 42c370b9-4569-d348-85c3-ab3aefc53057