Made the Player::selected a set of Entity pointers instead of the pointer to the head of the selected list. Entities no longer have a member to the next node in the selected list, too. The former data structure is simple, fast and portable between C and C++, but less straightforward and confusing when it comes to multiplayer. Entity selection may be something to do with the client only, but multiple players can select the same Entity, which do not cope with selection list.
This change is somewhat big, but I decided it was necessary. If, one day, we change mind back to think that selection list is more appropriate, we'd implement it with template instantiation. This is desirable way to develop in C++.

Also, temporarily disabled unserialization of Player's entity selection, to prevent the server from overwriting the client's selection. The context data should contain a flag indicating whether currently loading from a save file or unserializing data stream from the server.

git-svn-id: svn://localhost/gltestplus/branches/bullet@590 42c370b9-4569-d348-85c3-ab3aefc53057