Fixes #38 Changed the definition of the Serializable object id in serialized stream. Previously it was index of the objects in the stream. Now it's unique id in the server that is persistent throughout the simulation, namely Serializable::id.
Serialization method is totally changed in that protocol version must be incremented and game save / load function is temporarily unavailable.

The advantage of this change is that redundant pass in serialization is reduced, because all objects know it's id without accessing some other map. As such, SerializeContext's constructor no longer has an argument of reference to a map.

This step is necessary for making the stream compressable with run-length encoding, since all pointers have been converted to index in the stream, making the stream frequently alternate everytime some objects get deleted and the index shifts. Object's id, on the other hand, never changes once it is allocated, so it'd be suitable for compression with frame deltas.

git-svn-id: svn://localhost/gltestplus/branches/bullet@642 42c370b9-4569-d348-85c3-ab3aefc53057