Fixes #38 Stopped generating Serializable objects' id with any constructor. Instead, the creator must call one of the constructor with an argument of Game explicitly to generate unique id. Otherwise id is initialized with 0, which means an invalid object.
This distinction is necessary to indicate whether the object being created really born to world, or just unserializing from a stream. Game pointer is now a member of Serializable object to quickly refer the surrounding object.
Theoretically, it could be a reference, because whenever a Serializable object is being created, there's always a Game object, either ServerGame or ClientGame. We'd know if a Serializable object id should be fetched with determining Game type. But it was way too influencing everything that I gave up.

Made StellarFileLoad function a mamber of Game in the process.

Temporarily (or maybe permanently) disabled codes for Scarry, because it has been no more than burden of project build and code management with multiple inheritance. Now that Serializable object tree does not cope with multiple inherits, we'd seriously review the codes to bring Scarry back to life, which won't be a near future.

Added game variable to Squirrel envirionment, which cannot provide anything yet to Squirrel codes.

git-svn-id: svn://localhost/gltestplus/branches/bullet@643 42c370b9-4569-d348-85c3-ab3aefc53057