Refs #38 Added BinDiff class, the binary difference checker that returns difference of two given streams as a list of Patch objects. The returned Patch list is then be able to be applied to source stream to produce destination stream.
If the source stream and the destination is much alike, BinDiff will create quite few Patches that could be used for compressing stream updates.

Protocol major number is incremented because there're so much difference of serialization stream definition, but Linux dedicated server is not cared yet. This version cannot connect the server and the client at all, because the server needs to send the first stream without using diffs, which does not yet.

git-svn-id: svn://localhost/gltestplus/branches/bullet@646 42c370b9-4569-d348-85c3-ab3aefc53057