Fixes #38 Made the stream binary formatted. This will reduce the stream size to about a half.
In the process, we had to introduce individual operators to SerializeStream and UnserializeStream for char, unsigned char, short and unsigned short.
gltestp::dstring is added for unserialization destination, too.

Of course, protocol major version is incremented.

The main selection window in connected Windows client does not disappear for unknown reason.

git-svn-id: svn://localhost/gltestplus/branches/bullet@657 42c370b9-4569-d348-85c3-ab3aefc53057