Fixed the other players do not clear the pointer to an object being deleted in the server. It has not been exposed because the other players do not anim() either.
Also made the server correctly destroys Player object when the corresponding client socket is closed.

Added helper functions to inspect the Player's id from the Squirrel codes.

Probably it's the limit of the scheme to clear pointers to objects being deleted in postframe process. The observer scheme can be better solution to deal with dangling pointers.


git-svn-id: svn://localhost/gltestplus/branches/bullet@668 42c370b9-4569-d348-85c3-ab3aefc53057