Refs #40 Experimented with a test code that proves the current Observer scheme has a flaw. This revision should crash seconds after starting the game.
The Observer object itself must be a Observable and notify its death when gets deleted to Observed object, too. Otherwise, the observed object will have a dangling reference in the Observer list.
It sounds complicated, but to put it simply, the observer and the observed object must have references to each other. In this sense, there's no distinction between Observer and Observable.

The test code is set up to crash the program, but it won't always do, because the client could allocate another copy of Entity to the memory location the deleted object had been resided.

In the process of setting up a test, folloing changes are made.

* Added cmdlog.log output.
* Added Entity::enemy member visible to Squirrel codes.
* Made Entity inherit Observer interface too.
* Commented out postframe().

git-svn-id: svn://localhost/gltestplus/branches/bullet@674 42c370b9-4569-d348-85c3-ab3aefc53057