Refs #40 Changed the data structure of the Entity list in WarField to std::list of pointers. This is very tough decision indeed, because the former method had certain advantages that are not negligible.
The former data structure was the pointer chain. Each node had a pointer to the next node, and the root (container) object only had a pointer to the first node in the list. The last node's pointer is assigned NULL.
The major advantage of the pointer chain is its simplicity. It requires no additional memory usage other than the pointers (no separate heap blocks), and you can scan entire list in one direction in linear time.
The code to iterate the list looked pretty simple (compared to C++ STL's iterator idiom), and you can insert or delete a node in the list with a pointer-to-pointer like iterators.

The std::list data structure may be a bit too complex to store Entity list, for it requires additional heap memory for each node.
We will randomly delete the nodes from the list, and don't want the smart pointers in the list to reallocate so often, so we refused to use std::vector or std::deque.

Of course, searching in the pointer chain is not fast, but this data structure is used in where searching is rarely performed.

Another important property of the data structure is that multiple containers of the same type cannot contain the same node. This property is not suitable for arbitrary list, such as selection list, which could overlap among the selectors.
But again, the data structure is used to really contain objects inside something, meaning no duplicate containers will occur on the same node.

So far, I have described the advantages of the old method. Then why we bother to change it?
The main reason is that we use smart pointers (so-called weak pointers) to manage references to the nodes. Those pointers must know how to react to a object being deleted. We must dynamically change the method to clear the expired pointer for regular pointers (NULL assign) and the pointer chain's node pointers (retrieve the next node's next pointer).
But I didn't want each pointer to have a VFT, which would double the pointers' size.

I came to think that having VFT for each smart pointer is not that bad, but I want to commit the version without it for once.

I'm not yet decisive, but translating old codes that were depending on internal data structure won't hurt in the future.
At last, we could make a wrapper class for the (almost) pointer chain that resembles the syntax of std::list.

At least demo1.nut lasted for 30 minutes without errors as I write this message.

Writing long log messages for a revision like this is not common, but the range of influence is so large that one will lose sight of description in source comments.
So large I considered branching, but maintaining much like branches in parallel will cause problems.


git-svn-id: svn://localhost/gltestplus/branches/bullet@683 42c370b9-4569-d348-85c3-ab3aefc53057