Changed the base type of Player::selected and Player::chases from std::set<Entity*> to std::set<WeakPtr<Entity>> to respond to Entity deletion.
It's not really good to have a set's element value altered after insertion, because the set container works in an assumption that all elements have consistent ordering, which could corrupt if altering is allowed.
The set container itself should respond to contained object's deletion (as Observer or something), in order to keep all contained pointers valid and consistent.

But for the moment, it works.
Former bug that crashed the game when a selected Entity is destroyed is fixed, and the game ran stable for 10 min.

git-svn-id: svn://localhost/gltestplus/branches/bullet@693 42c370b9-4569-d348-85c3-ab3aefc53057