Got rid of many obsolete functions, namely sqa_refobj, sqa_newobj, sqa_set and sqa_get. It becomes much simpler.
There were many codes that have been depending on those obsolete functions even after previous commits.
The best way is to delete them, and we've done it.

Made Docker a Observable. Probably Docker is not appropriate to be a base class of Observable tree, but the existing code suggested doing so.

git-svn-id: svn://localhost/gltestplus/branches/bullet@701 42c370b9-4569-d348-85c3-ab3aefc53057