* Separated ClientMessage class definition from Applicaion.h to ClientMessage.h and made the class EXPORTed. * Modified EXPORT macro definition conditionals to prevent macro expansion from messing up Doxygen parser. * Added CMDockCommand to enable the client to request the server that the player wants an Entity begins docking. The server can refuse the request if the client does not have right privilege. * Added a mechanism to enable Squirrel codes to sendClientMessage. * Made ServerGame::idmap() have mapping of objects and SerializableIds just like the client.
I've considered making ClientMessage and EntityCommand inherit one way or the other, because they have similar mechanisms and similar purposes. But I came to think that they are distinct in that EntityCommand always take effect, while ClientMessage can be ignored if the server decides so. EntityCommand does not contain information about the commander, so it cannot mimic the behavior of ClientMessage easily.
Most EntityCommands also can be sent as ClientMessages, such as CMMove and CMDockCommand, but it does not mean EntityCommand is same as ClientMessage or subset or superset.

git-svn-id: svn://localhost/gltestplus/branches/bullet@705 42c370b9-4569-d348-85c3-ab3aefc53057