* Added CMEntityCommand to send (almost) all EntityCommands without duplicating classes. The EntityCommands that allow sending over the network must inherit SerializableCommand class. It makes the object serializable in a similar way as Serializable class, but not equal; SerializableCommand does not have object id. The EntityCommandStatic record's unused newproc function pointer is activated in order to dynamically create EntityCommand-derived objects based on class names. This mechanism is only demonstrated in RemainDockedCommand, but other commands should make use of it too.
* Fixed astro.cpp wasn't included in the build target in VC.

* Fixed ClientMessage.h couldn't be included on its own. See the issue below.

Fixed #41
* Worked around a silly issue of VC2008 that causes duplicate definition in Winsock.h and Winsock2.h if Windows.h is included before Winsock2.h, by defining _WINSOCKAPI_ macro so that winsock.h won't be included.
I'm not sure if it's the best way to work around, but at least it only affects VC project files. If later versions of VC resolves this issue or change the macro name, we could revise this issue.
I want to place the include directives for custom classes to the top, because it's the only way to ensure the header includes all necessary headers. That Microsoft issue disallow me to do it when the inclusion dependency graph gets complicated, mixing Windows.h in the way.

git-svn-id: svn://localhost/gltestplus/branches/bullet@717 42c370b9-4569-d348-85c3-ab3aefc53057