Made Beamer dockable to Docker in the docking bay (back of the structure in case of Shipyard). Shipyard also opens doors for docking Frigates now.
There's an issue that assertion fails in Bullet dynamics engine when undocking a Beamer just after it is docked. Probably leaveField is not called properly since WarField::endFrame() is no longer called.
Another wonderful issue is that Beamer seems to colliding something that doesn't appear when docking to Shipyard.

In the process, the following changes are done.

* Docker::getPortPos() and Docker::getPortRot() now have an Entity argument that can be used to change port position depending on actual Entity being docked.
* Added dockque() to Docker member functions to handle the event when an Entity is being docked.

git-svn-id: svn://localhost/gltestplus/branches/bullet@730 42c370b9-4569-d348-85c3-ab3aefc53057