Made MQO files and motion files in a zip file (rc.zip) can be loaded.
In the process, mqo.h and mqo.cpp is generalized in that input stream can be specified with std::istream. The source is no longer necessarily a disk file.
In exchange, a C source can no longer easily call the functions. In the end, Model and Motion classes won't be utilized from C. We could give up supporting C at all.


git-svn-id: svn://localhost/gltestplus/branches/bullet@735 42c370b9-4569-d348-85c3-ab3aefc53057