Experimentally enabled Shipyard's overlay icon customizable in the Squirrel initialize script (Shipyard.nut). I've tried raw object handling functions of Squirrel library (sq_getstackobj, sq_addref, etc) and OpenGL display lists, and currently accepted the latter. The former did not work well when the initialized VM (in this case, server) and applied VM (client) is different. The documentation of Squirrel does not clearly states whether inter-VM object handling is supported, but I feel not.
Making the icon customizable through scripts is fun, but may not much of use.

Also virtualized arguments of Warpable::sq_init() to enable arbitrary parameters to be specified. It can be some extra load, for it uses heap memory via std::vector of pointers, but it'd be minor as long the initialization code is only executed once in a process lifetime.

git-svn-id: svn://localhost/gltestplus/branches/bullet@739 42c370b9-4569-d348-85c3-ab3aefc53057