Made the edit window only visible when focused. It's hidden (in term of Windows window system) when not focused, so it no longer catches mouse and key events.
The pinned window can be focused for now.

focusEnter() and focusLeave() event handlers are added to GLwindow to make it possible.
In the process, glwfocus is made a static member of GLwindow, which requires calling getFocus() to retrieve the value from outside.
Whenever focused GLwindow is to be changed, invocation of focusEnter() is necessary. Likewise, defocusing comes with focusLeave(). Normally this will be assured by focus() and defocus().

Also fixed Assault didn't load hitboxes properly.
Also made mover console command available to dedicated client.

Stable for 10 min.

git-svn-id: svn://localhost/gltestplus/branches/bullet@756 42c370b9-4569-d348-85c3-ab3aefc53057