Made Entity::controller a WeakPtr<EntityController> to manage object life. Also added it to the serialization target.
In most (at this time, only) cases, EntityController is just a Player, so the moment the controlling Player logs off is when the pointer must be cleared.

Made Assault updates per frame in the client game too. As a result, it will move smoothly in the client and changes color of engine blast (if fragment shader is enabled).

Also made Assault::draw() use Model class instead of old suf object interface, while keeping the shader modulation enabled. Cleaned up old codes and comments as well.

Modified AdditiveShaderBind::setIntensity() so that debugging with IDE is a bit easier.


git-svn-id: svn://localhost/gltestplus/branches/bullet@760 42c370b9-4569-d348-85c3-ab3aefc53057