The previous commit did not really fix the problem, it just removed a btDynamicsBody from the Entity going to warp.
The real problem is that Entity::unserialize() occasionally calls enterField().
If the newly assigned WarField is just created in the same frame, the WarField object in the client can be yet uninitialized at the time of unserialize() call, because it's the function WarField object will be initialized too and the order of unserialize() calls is not determined. (In effect, it's the order of creation, i.e. oldest first.)
We resolved it by adding a special flag to Entity indicating the object is just changed containing WarField, to defer execution of enterField() from unserialize() to callClientUpdate().

Entity::unserialize() can call leaveField() too, but leaveField() can never be called with uninitialized WarField (yet).

Lesson here: do not put processes that depend on contents of other Serializable objects to unserialize(). If it's necessary, defer it to next anim() or clientUpdate() call.

git-svn-id: svn://localhost/gltestplus/branches/bullet@786 42c370b9-4569-d348-85c3-ab3aefc53057