Replaced Initializator defined in astrodraw.h with StaticInitializer and made a dedicated header for it.
The previous class Initializator was introduced for drawing methods, but it's so generic that nond-drawing methods can make use of it. So we decided to make it non-dependent of any other modules in the project. In fact, it could be a part of cpplib.

As a demonstration, Game class's serverInits list (that is encapsulated in a function with construct on first use idiom) gets filled with StaticInitializer objects. 

Game::clientInits() is added for cmd_warp to be registered in Warpable.cpp, too.

git-svn-id: svn://localhost/gltestplus/branches/bullet@788 42c370b9-4569-d348-85c3-ab3aefc53057