Revived weapon type, ammo count readings and crosshair to the HUD of Soldier.
Also fixed a bug orientation of an Entity's field of view reset to identity when the player tries to control the Entity. It involved the server and the client and it required both to initialize the Player's orientation correctly.

Also fixed an Entity in the client thinks it's still being controlled even after the Player uncontrolled it.

git-svn-id: svn://localhost/gltestplus/branches/bullet@868 42c370b9-4569-d348-85c3-ab3aefc53057