Made starting control always capture the mouse in the client.
Also experimenting with GLWcontrolButton if the class can be altered with Squirrel callbacks (GLWsqStateButton).
The results show GLWsqStateButton implementation is significantly slower than GLWcontrolButton, as I expected. It took 1e-5 sec to query the value of button state from the game. It doesn't look much, but it's about 1600th of a frame time in 60 FPS, the button is always shown, and experimented CPU may be faster than actual execution environment.
As a reference, GLWcontrolButton took 1e-8 sec. It's almost as short as detection limit.

The obvious thing is that cmd_control console command need not to be defined in C++ source but a Squirrel function.

git-svn-id: svn://localhost/gltestplus/branches/bullet@894 42c370b9-4569-d348-85c3-ab3aefc53057