Brought back mission script selecting ability, which has been disabled since the network game is introduced.
Internally, we made CMSQ function in Squirrel script accepts an argument to run arbitrary command in the server.
It's dangerous to run arbitrary code or script file from the client, so we added protection layer in common.nut which made the system a bit complicated.

Only demo missions are handled in this revision, and all mission scripts with initUI() don't work.

frameproc() in mission scripts does not work in this revision for some reason.

git-svn-id: svn://localhost/gltestplus/branches/bullet@922 42c370b9-4569-d348-85c3-ab3aefc53057