Now that fuzzy file names are disallowed for textures since r903, Defender and Sceptor have been without textures on them. Defender was fixed to draw the model with LoadMQOModel() and DrawMQOPose(), but Sceptor is not yet.
It's redundant and will consume slightly more memory and CPU time to have symmetric motions defined in motion files, but things are more customizable now.
Disappointing thing is that the redundant model will be tougher to modify.

TODO: we would be better to use the MotionPose object to obtain the leg position, rather than calculating with a hard-coded logic.

git-svn-id: svn://localhost/gltestplus/branches/bullet@968 42c370b9-4569-d348-85c3-ab3aefc53057