Changed GLWchart::TimeChartSeries' label string to most recent value instead of maximum value in range.
Range of the values is suitable for axis label, but what we actually want to
know is the current value, not the maximum of last few values.
It shows the difference when several TimeChartSeries'es share the same ygroup
but have very different dynamic range of vaules. Namely, very short process time
versus the full frame time.

git-svn-id: svn://localhost/gltestplus/branches/bullet@1012 42c370b9-4569-d348-85c3-ab3aefc53057