Made Soldier's hook attach to Entity's parts that move relative to the attached Entity's orientation.
TorusStation demonstrates this situation.
The station have rotating wheels (that are fixed to TorusStation's orientation)
and non-rotating ports at both ends of the hub (that are revolving in opposite
direction of the wheels).  If you attach the hook to the wheels, you would
follow the wheels' rotation.  If you hit the ports though, it seems stationary.

Also made Soldier.cpp's line endings consistent CRLF, which may create a large
diff.

git-svn-id: svn://localhost/gltestplus/branches/bullet@1037 42c370b9-4569-d348-85c3-ab3aefc53057