Derived GimbalTurret to create a new turret type that shoots Missiles. No appropriate models are designed yet.
The Entity::EntityRegister class template seems to had a bug that caused 
compile error in cascaded inheritance such as GimbalTurret and 
MissileGimbalTurret.
I was surprised that there have been no such case until now, but it seems we 
have overlooked this bug for quite some time.
Just making accessibility properties appropriate for member functions solve
this problem.

It seems that Missile::MissileList has a bug that causes runtime error when
the program terminates with some Missiles alive in the game space.
We didn't fix it in this revision, but we would probably be better making
the list a plain STL container.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1065 42c370b9-4569-d348-85c3-ab3aefc53057