Experimentally added function to predict Entity creation in the client.
The objects created in the client are deleted once the update stream is 
received from the server, because the server will tell us if the predicted 
objects are really created in the server too.

We changed the updateInterval value temporarily longer to see the effect of 
prediction. It seems a bit inaccurate but works. Maybe it's time to make 
frame rate fixed.
One point made me disappointed is that Missile trail smokes disconnect when
an server update is invoked.

Also made Game.cpp's line endings consistent LF. Entity.cpp is made CRLF.

git-svn-id: svn://localhost/gltestplus/branches/bullet@1069 42c370b9-4569-d348-85c3-ab3aefc53057