Added Teline3D and Tefpol3D's debugging output to GLchart.
We suspected that there was a leak of vertices for Tefpol3D.
So we added the debugging info to the chart and found out
"teplcount" never decreases, capped at 128.

It only works in debug build.

We finally added GLWchart::addSamplesToChart() that many codes
write the same codes in place. The "tellcount", "teplcount"
and "tevertcount" are the first ones that demonstrates the
functions.

git-svn-id: svn://localhost/gltestplus/branches/bullet@1087 42c370b9-4569-d348-85c3-ab3aefc53057