Prevented Tefpol leak by adding destructors to BeamProjectile and Sceptor classes.
Also we made sure that Tefpols to be created only in the client by adding
game->isClient() to conditionals.
Note that isClient() is not equivalent to !isServer().  We could be making
standalone game, which is a client and a server at the same time.

Of course, there must be other cases that leak Tefpols since network game 
is introduced.  They should be fixed shortly.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1090 42c370b9-4569-d348-85c3-ab3aefc53057