Added active flag to Bullet class to prevent it from passing through objects in the client's motion prediction.
If a Bullet is determined to hit some object, that Bullet won't travel 
further nor be drawn.
In this way, we won't see hit effect several times for a bullet.

Also made Bullet's member variables protected and added getter appropriately.
Many codes assign Bullet::life directly, but we can pass it to the 
constructor of Bullet, which is better way to do it.

Now, it looks more like standalone game even if it's network game.

git-svn-id: svn://localhost/gltestplus/branches/bullet@1098 42c370b9-4569-d348-85c3-ab3aefc53057