Renamed Entity::bullethit() to onBulletHit() and revived Frigates' shields.
The function onBulletHit() has an additional hitpart argument to tell the 
object the part bullet hits.
We use this value to determine whether the shield should be damaged instead
of ship hull.
The hitpart value is the same as return value of Entity::tracehit().

Also made Frigate predict motion in the client by default.

All the improvements are demonstrated in demo2.nut.

git-svn-id: svn://localhost/gltestplus/branches/bullet@1099 42c370b9-4569-d348-85c3-ab3aefc53057