Revived tutorial1 to some extent.
The tutorial mission can be played on the network game, although I don't 
see a reason to play a tutorial mission on the net.

The progress could be controlled with CMSQ (Client Message from SQuirrel), 
but it won't show message texts in the client if the server is remote 
computer.

The funny thing is that tutorial messages are in English even if the
client's UI strings are set to Japanese, because the language setting
can diverse between server and client.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1100 42c370b9-4569-d348-85c3-ab3aefc53057