Included Mersenne Twister to the library archive in gcc Makefile. Also added license terms file.
rseq_test.c covers the performance test for Mersenne Twister with -m option.

I'm still dubious whether to include Mersenne Twister to my library 
because of licensing issues, but I cannot evaluate it without including it.

As I thought, Mersenne Twister is so slow to initialize compared to simple 
linear congruential generators, which means it's not suitable for my use.

On the other hand, generation time of each random number is comparable to
LCG.


git-svn-id: svn://localhost/projects/clib@202 c24b926b-5ef4-9a42-8c70-094e0699c819