Added Defender to Builder's recipe set.
We pass the recipe name (in this case, "Sceptor" or "Defender") to specify 
what to build.  This method should be most extensible, but may not be 
efficient compared to passing recipe index.  I won't bother profiling, 
though.  Though if it's inefficient, it just alters once in a very while 
(more than several seconds), which means optimizing the process is not much
beneficial.

Also fixed progress bar's Y position in GLWbuild.

git-svn-id: svn://localhost/gltestplus/branches/bullet@1133 42c370b9-4569-d348-85c3-ab3aefc53057