Made Builder::ru accessible from Squirrel codes.
As a demonstration, demo1.nut's Shipyards gain resources periodically.
It's good to read from or write to the value while building something.

Maybe explicit getter and setter (or maybe adder) closures are preferable, 
but for the time being, it works by delegating the whole process from 
sqf_Entity_get() and sqf_Entity_set() to Builder::sq_get() and sq_set().


git-svn-id: svn://localhost/gltestplus/branches/bullet@1134 42c370b9-4569-d348-85c3-ab3aefc53057