Fixed the bug that caused segmentation fault when a Sceptor is destroyed after its mother Shipyard is destroyed.
We haven't fully implement server and client's weak referencing, due to lack 
of understanding and testing.  But this commit is a good starting point that
inform us that what type of errors can occur.

The script demo1 also tests server side added Entities to a ShipyardDocker,
which can be treated in slightly different way from Bilder-constructed
Entities, and it works.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1168 42c370b9-4569-d348-85c3-ab3aefc53057