Experimentally added the task bar on the left edge of the game screen.
A new class GLWtaskBar is added for this purpose.
Chat window button is shown in the task bar as a demonstration.

Also allowed transparent (textures with alpha channel) in
GLWcommandButton::draw() graphics.
This is especially useful for icons that are authored with Inkscape,
because it can output PNG files with alpha channel that can be read
directly by the program.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1169 42c370b9-4569-d348-85c3-ab3aefc53057