Moved Build Manager and Dock Manager buttons from Cameras button matrix to the task bar.
Also made these buttons to show status whether they can be pressed
right now (namely, if selected list of Entities contain Builders or Dockers).

In this process, we allowed GLWstateButton to have null as the second
argument to the constructor in Squirrel scripts.  If so, the disabled 
(inactive) state is shown as darkened version of enabled (active) state's
button image.  We can skip efforts to create active/inactive pairs for all
state buttons.

Also made "buildmenu" command to show the Build Manager window if any one
of selected Entities is a Builder, not just the first one.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1172 42c370b9-4569-d348-85c3-ab3aefc53057