Made the tool tip window (shown when mouse cursor is hovering over a button) keep in the screen even if the button is at the top of the task bar.
Technically, the GLWtip window is shown inside the screen (OpenGL viewport)
even if it's near the top or left edge.

Also made the tool tip window insensible to mouse inputs and pass the mouse
events to the window under it.

In this process, we added mouseHit() virtual function to customize hit
shape of GLwindow-derived classes.  In GLWtip's case, it never hits mouse.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1173 42c370b9-4569-d348-85c3-ab3aefc53057