Made GLwindow system's constructors have mandatory Game argument.
Keeping the obsolete version of constructor (without Game argument) may be
dangerous because Squirrel scripts assume all in-game Serializable objects
have game member variable.  We will be required to type extra keys for the
new rule, but I believe that forcing the rule at this time will keep us from
subtle problems in the future.
In reality, we have always the same Game object (application.clientGame) for
GLwindows, while the original intent of Serializable::Game is to support
multiple instances of Game objects in a process.

Also made window title strings dstring.  It was created so long before that
dstring was not stable, but now we don't want to manage heap memory for
strings manually.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1175 42c370b9-4569-d348-85c3-ab3aefc53057