The first commit for trying to introduce standalone game mode.
The standalone game mode here means that there's no server thread and
application's serverGame and clientGame points to the same object.

The mode is desirable for tutorials, which prints messages and such
instructions to the player's screen frequently.

We got rid of isClient member variable from Application because the game
mode came to be so complex that single isClient flag is insufficient to
decide anything.
Application::mode now takes the role of isClient.

Also cleaned up old codes that assumed there's a waiting room before game
starts.

Many inter-game messages haven't supported the feature, so it's still buggy
and crashes after moments from window's appearance, but we need some
starting point to do something.
In fact, it could deserve a branch, but we haven't branched when we
introduced network game itself, so it's a bit off balance to branch now.

The bottom line is that previous game modes (ServerGame and ClientGame)
works.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1186 42c370b9-4569-d348-85c3-ab3aefc53057