Tried a very straightforward way to define message sequence with Squirrel generators in tutorial2.nut.
If we want to modify the message sequence by a method based on the player's
actions, adding closures to determine next state is one way, but it's
cumbersome to maintain if the sequence and condition get complex.
Generators or coroutines are handy in such a case, although I haven't been
confident on them until this revision's demonstration.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1201 42c370b9-4569-d348-85c3-ab3aefc53057