Tried thread (coroutine) of Squirrel instead of generator.
It works, but is less straightforward and brings in subtle issues which
can only be resolved by defining custom native closure in place of
Squirrel's default newthread().

It has benefit over generators that enables the coroutine to call another
function that has suspend() invocation.
In a generator, you can 'yield' only in the generator's main function.
But I feel it doesn't worth the effort.

Also made GLWmessage not to call uninitialized destroy callback for Squirrel.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1202 42c370b9-4569-d348-85c3-ab3aefc53057