Made tutorial3.nut to detect if the player opens the build manager window.
Previously, the build manager window generation is hard-coded into the
executable, so we could only invoke "buildmenu" console command to show one.
But now, many other window classes switched to more script-oriented way, so
we follow the custom and return window handle to the script so we could
use it later.


git-svn-id: svn://localhost/gltestplus/branches/bullet@1208 42c370b9-4569-d348-85c3-ab3aefc53057