Switched to third-party resource editor for authoring .rc file.
This is because Visual Studio 2012 Express does not come with
the resource editor, which is needed to create or customize initial
settings dialog with a GUI editor.
For the sake of portability, I don't feel like relying on a specific
platform's GUI development environment, but I cannot help using
resource script editor for Windows right now.

At least we can use freeware based workflow.  ResEdit is such a
freeware that I ported the existing resource script to in this commit.

Probably we could use open source GUI framework such as Qt.