Moved Squirrel build target to local subdir for GNUmakefile.
A mkdir command is added to let a clean git checkout build the target
libraries.  We don't really need sq command line tool, but the default
Squirrel Makefile does not provide a way to avoid it's creation.
We may be permitted to modify the Makefile in the Squirrel source
according to squirrel3/COPYRIGHT, but I'm afraid doing so will make
future merge with library updates hard.