Added the Earth surface textures to source control.
Also updated credits.txt so that it denotes sources of the textures.
The EOL style of credits.txt is also changed to LF.

Unlike human-crafted images (such as button icons), planetary images
are large even if compressed.  The Earth's night lights texture was
surprisingly hard to compress without losing details.