Fixed scaling of TIN model for ASTER GDEM.
The ASTER GDEM data tile is a degree square.  The latitude edge should
have length of 111 kilometers, but apparently it looks better with
the length halved, i.e. 55 kilometers.

The triangulated GDEM data is not included in the version control,
because it's large data (4 megabytes).
If you want to test it, just download a sample data by running

	wget https://www.dropbox.com/s/5epztg1ioy7fgl4/triangles.all

in the root directory.