surface: Sped up loading TIN and its resampled DEM.
TIN-resampled DEM is cached into cache subdirectory at the first
invocation of the program.
Repeated startup of the program loads the cached data to shortcut
the calculation needed to resample TIN.

This workaround is not really a right way.  We should enhance
the algorithm to quickly detect a triangle that contains a given
point.  But it's known to work and really effective (except for
the first run, of course).